Use interleaved MEX API, clean up MATLAB installation (#18)
* Use interleaved complex MEX API
* Clean up MATLAB installation, add configuration script
* Update examples, install script, notes

+++ src/libstoast/pparse.cc
@@ -59,7 +59,7 @@ void ParamParser::Lineout (const char *str)
 bool ParamParser::GetString (const char *cat, char *str)
 {
     char cbuf[256];
-    int i, ok, lcat = strlen(cat);
+    int i, ok, lcat = (int) strlen(cat);
     
     if (!bfile) return false;
     prmf.clear();
